requirements.txt
streamlit
google-cloud-firestore
pandas
python-dotenv
toml